# perf: ~45% faster Desk first response