--- a/frappe/website/path_resolver.py
+++ b/frappe/website/path_resolver.py
@@ -17,6 +17,8 @@ from frappe.website.utils import can_cache, get_home_page
 
 
 class PathResolver:
+	__slots__ = ("path",)
+
 	def __init__(self, path):
 		self.path = path.strip("/ ")
 
@@ -36,6 +38,11 @@ class PathResolver:
 			return frappe.flags.redirect_location, RedirectPage(self.path)
 
 		endpoint = resolve_path(self.path)
+
+		# WARN: Hardcoded for better performance
+		if endpoint == "app":
+			return endpoint, TemplatePage(endpoint, 200)
+
 		custom_renderers = self.get_custom_page_renderers()
 		renderers = custom_renderers + [
 			StaticPage,