--- a/frappe/website/path_resolver.py
+++ b/frappe/website/path_resolver.py
@@ -105,7 +105,7 @@ def resolve_redirect(path, query_string=None):
 	        ]
 	"""
 	redirects = frappe.get_hooks("website_redirects")
-	redirects += frappe.db.get_all("Website Route Redirect", ["source", "target"])
+	redirects += frappe.get_all("Website Route Redirect", ["source", "target"], order_by=None)
 
 	if not redirects:
 		return