Major changes:

- Hardcode `/app/` for higher resolution priority. Currently everytime you open Desk whole lot of crap is checked before just rendering it as template. This includes web forms, static pages, web views etc etc. Just utter waste. 
- Consistently use Website settings from cache only. Half the time it's used from cache, other time it was from DB which meant it was worst of both worlds. Found all usage in critical path and converted all to cached version.
- Remove preload header computation and automatically add any included bundled assets to preload instead. Currently, we are parsing responses to find css/js files... which in itself was overhead and required hacks like straining HTML before parsing. 😄 
- Use cached `is_virtual_doctype` function
- Drop `order_by` from queries where it doesn't matter. 

TODO:
- [x] ~~remove preload hints completely?~~ This was added here https://github.com/frappe/frappe/pull/10229 but it's not in use on FC https://github.com/frappe/agent/commit/24d08328545d0249afd8aa6e196dbda3e480b635 Though not useful for server push much, preload hints still help. Current implementation has almost zero overhead so 🤷 



#### Numbers - TTFB measured using cURL (n = 100)

This drops `/app` query count from 33 to 8! (check using recorder)


  | Before | After | Improvement
-- | -- | -- | --
p50 (median) | 0.156 | 0.084 | 46.1%
p75 | 0.158 | 0.087 | 44.7%
p90 | 0.165 | 0.091 | 44.9%


In simple words, ~**70 milliseconds** fewer on each desk load. 🤌 

<details>
  <summary>Test setup and script</summary>

- Disable developer mode
- Set `frappe._dev_server = 0`
- Use gunicorn instead of werkzeug: `gunicorn -b 127.0.0.1:8000 -w 1 -t 120 --chdir /Users/ankush/benches/develop/sites frappe.app:application --preload`
- copy /app request from browser as cURL. Modify it a bit like this:

```bash
repeat 100 sleep 0.4 &&
curl -o /dev/null -s -w '%{time_starttransfer}\n' 'http://develop:8000/app' \
  -H 'Accept: text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9' \
  -H 'Accept-Language: en-GB,en-US;q=0.9,en;q=0.8' \
  -H 'Cache-Control: max-age=0' \
  -H 'Connection: keep-alive' \
  -H 'Cookie: sid=0081c74af96540c1ef032a46cc296e053727a0ec28bd87d6131799d8; system_user=yes; full_name=User%20C; user_id=lowperm%40user.com; user_image=' \
  -H 'Referer: http://develop:8000/' \
  -H 'Upgrade-Insecure-Requests: 1' \
  -H 'User-Agent: Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/104.0.0.0 Safari/537.36' \
  --compressed \
  --insecure >> new.txt

```
</details>

closes https://github.com/frappe/frappe/issues/17885 caused by https://github.com/frappe/frappe/pull/17785 
closes https://github.com/frappe/frappe/issues/17001 
closes https://github.com/frappe/frappe/issues/17423