@sagarvora 2022-08-21
> * Drop `order_by` from queries where it doesn't matter.

Is it a bad design in terms of performance to default `order by` to something?