@sagarvora 2022-08-21
One other place where I think there are a lot of possibilities to speed this up is `boot.py`