@ankush 2022-08-21
> Is it a bad design in terms of performance to default order by to something?

Maybe, it's sane for most other usage outside of core framework I guess 😄 

I removed them cause I noticed it randomly, unnecessary order by just causes _little bit_ of confusion to the query planner _sometimes_. It ends up picking an index that doesn't contain the column being fetched. Very _micro-optimziation_ tbh.