@sagarvora 2022-08-21
> * Hardcode `/app/` for higher resolution priority. Currently everytime you open Desk whole lot of crap is checked before just rendering it as template. This includes web forms, static pages, web views etc etc. Just utter waste.

This is a great optimisation. People open tabs of desk way too often as compared to anything else.